# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.ParkingSpace:
    properties:
      parkingLotID:
        type: integer
      spaceID:
        type: integer
      status:
        type: string
    type: object
  models.PaymentRecord:
    properties:
      amount:
        type: number
      paymentID:
        type: integer
      paymentMethod:
        type: string
      paymentTimestamp:
        type: string
      recordID:
        type: integer
      reservationID:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: 1351688749@qq.com
    name: API Support
    url: http://example.com/support
  description: API documentation for Parking Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Parking Management API
  version: "1.0"
paths:
...
  /api/parkingspace:
    post:
      consumes:
      - application/json
      description: Create a new parking space
      parameters:
      - description: Parking Space Details
        in: body
        name: parkingSpace
        required: true
        schema:
          $ref: '#/definitions/models.ParkingSpace'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a parking space
      tags:
      - ParkingSpace
  /api/parkingspace/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the parking spaces by parking lot id
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ParkingSpace'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking spaces by parking lot id
      tags:
      - ParkingSpace
  /api/parkingspace/{lotid}/{spaceid}:
    put:
      consumes:
      - application/json
      description: Update the status of a parking space
      parameters:
      - description: Parking Lot ID
        in: path
        name: lotid
        required: true
        type: integer
      - description: Parking Space ID
        in: path
        name: spaceid
        required: true
        type: integer
      - description: Updated Parking Space Details
        in: body
        name: parkingSpace
        required: true
        schema:
          $ref: '#/definitions/models.ParkingSpace'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update parking space status
      tags:
      - ParkingSpace
  /api/parkingspace/free:
    get:
      consumes:
      - application/json
      description: Retrieve the free parking spaces
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ParkingSpace'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get free parking spaces
      tags:
      - ParkingSpace
  /api/parkingspace/plate/{plateNumber}:
    get:
      consumes:
      - application/json
      description: Retrieve the parking space by license plate
      parameters:
      - description: License Plate Number
        in: path
        name: plateNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ParkingSpace'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking space by license plate
      tags:
      - ParkingSpace
  /api/parkingspace/status:
    get:
      consumes:
      - application/json
      description: Retrieve the status of a parking space by id
      parameters:
      - description: Parking Lot ID
        in: query
        name: lotid
        required: true
        type: integer
      - description: Parking Space ID
        in: query
        name: spaceid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking space status by id
      tags:
      - ParkingSpace
  /api/parkingspace/user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the parking space by user id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ParkingSpace'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking space by user id
      tags:
      - ParkingSpace
  /api/paymentrecord/plate/{plate}:
    get:
      consumes:
      - application/json
      description: Retrieve the payment amount by plate number
      parameters:
      - description: Plate Number
        in: path
        name: plate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment amount by plate number
      tags:
      - PaymentRecord
  /api/paymentrecord/plate/{plate}/info:
    get:
      consumes:
      - application/json
      description: Retrieve the payment info by plate number
      parameters:
      - description: Plate Number
        in: path
        name: plate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment info by plate number
      tags:
      - PaymentRecord
  /api/paymentrecord/record/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the payment amount by record ID
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment amount by record ID
      tags:
      - PaymentRecord
  /api/paymentrecord/record/{id}/status:
    get:
      consumes:
      - application/json
      description: Retrieve the payment status by record ID
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment status by record ID
      tags:
      - PaymentRecord
  /api/paymentrecord/reservation/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the payment amount by reservation ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment amount by reservation ID
      tags:
      - PaymentRecord
  /api/paymentrecord/reservation/{id}/status:
    get:
      consumes:
      - application/json
      description: Retrieve the payment status by reservation ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment status by reservation ID
      tags:
      - PaymentRecord
...
swagger: "2.0"